PlotProp {BiostatTools}	R Documentation

PlotProp title

Description

Print plots

Usage

PlotProp(x, y, xlab = "", ylab = "", legend = "", main = "",
    method = "sisonglaz", stacked = F, percent = T,
    colors = c("grey", "orange", "lightblue", "pink"))

Arguments

x
A factor variable in a data.frame.

y
A factor variable in a data.frame. This will be used as groups and percentages of variable x will be computed for each level of y.

xlab
A character label for the x axis.

ylab
A character label for the y axis.

legend
A character label for the variable y.

main
A character label for the title othe plot.

method

stacked
If TRUE, a stacked bar plot is obtaind. Default is FALSE.

percent
If TRUE (default), percent values are printed.

colors
A vector with color names to be used in bars. Default is c("grey", "orange", "lightblue", "pink")

Examples

birthwt %>% with(PlotProp(factor(race),factor(ht)))
birthwt %>% with(PlotProp(factor(race),factor(ht)), stacked=T)
birthwt %>% with(PlotProp(factor(race),factor(ht)), stacked=T) + coord_flip()

## The function is currently defined as
function (x, y, xlab = "", ylab = "", legend = "", main = "",
    method = "sisonglaz", stacked = F, percent = T, colors = c("grey",
        "orange", "lightblue", "pink"))
{
    if (class(x) != "factor")
        return(paste("Sorry, the first variable is not a factor. This function requires two factor variables"))
    if (class(y) != "factor")
        return(paste("Sorry, the second variable is not a factor. This function requires two factor variables"))
    df <- table(x, y)
    res <- TablePropCI(df, method)
    if (percent & stacked) {
        p <- geom_text(aes(label = paste(round(est * 100, 2),
            "%")), stat = "identity", position = position_fill(vjust = 0.5))
    }
    else if (percent & stacked == F) {
        p <- geom_text(aes(label = paste(est * 100, "%")), position = position_dodge(width = 0.9),
            vjust = -0.25)
    }
    else {
        p <- NULL
    }
    if (stacked) {
        ggplot(res, aes(x = Cols, fill = Rows, est)) + geom_bar(stat = "identity",
            color = "black") + scale_fill_manual(values = colors) +
            xlab(xlab) + ylab(ylab) + labs(fill = legend) + ylim(0,
            1) + ggtitle(main) + p
    }
    else {
        ggplot(res, aes(x = Cols, fill = Rows, est)) + geom_bar(stat = "identity",
            position = position_dodge(), color = "black") + scale_fill_manual(values = colors) +
            xlab(xlab) + ylab(ylab) + labs(fill = legend) + ylim(0,
            1) + ggtitle(main) + p
    }
  }

[Package BiostatTools version 0.1.2 Index]