TableRR {BiostatTools}	R Documentation

TableRR

Description

Compute a risk ratio table.

Usage

TableRR(x, y)

Arguments

x
A variable of exposed groups.

y
A variable of events.

Details

The risk ration is computed assuming a cohort study defined by groups in x. The risk is computed for the second level of y.

Value

A formated risk ratio table

Examples

##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.

## The function is currently defined as
function (x, y) 
{
    df <- table(y, x)
    res <- riskratio(df)$measure %>% round(2)
    res %>% kbl(caption = "Risk ratio") %>% kable_styling(full_width = F, 
        html_font = "Cambria")
  }

[Package BiostatTools version 0.1.2 Index]